wip: general location spec for arrays also
