+++ src/earthkit/hydro/distance/_toplevel.py
@@ -1,36 +1,16 @@
 from functools import wraps
 
-import numpy as np
-import xarray as xr
-
 import earthkit.hydro.distance.array._operations as array
 from earthkit.hydro._backends.find import get_array_backend
 from earthkit.hydro._utils.decorators import xarray
-from earthkit.hydro.catchments.array._operations import preprocess_stations
+from earthkit.hydro.catchments.array._utils import locations_to_1d
 
 
 def _convert_locations(func):
     @wraps(func)
     def wrapper(river_network, field, locations, *args, **kwargs):
         xp = get_array_backend(river_network.groups[0])
-        orig_locations = locations
-        xr_present = isinstance(field, (xr.DataArray, xr.Dataset))
-        dict_locations = isinstance(locations, dict)
-        if dict_locations:
-            assert xr_present
-
-            lats = field.lat.data
-            lons = field.lon.data
-
-            # TODO: decide if acceptable to always use np here
-            locations = []
-            for lat_val, lon_val in orig_locations.values():
-                ilat = np.abs(lats - lat_val).argmin()
-                ilon = np.abs(lons - lon_val).argmin()
-                locations.append((int(ilat), int(ilon)))
-
-        locations = xp.asarray(locations, device=river_network.device)
-        stations_1d = preprocess_stations(xp, river_network, locations)
+        stations_1d, _, _ = locations_to_1d(xp, river_network, locations)
 
         result = func(river_network, field, stations_1d, *args, **kwargs)
